requirements.txt
streamlit
PyPDF2
scikit-learn
nltk
spacy
textstat
plotly
pandas
numpy
networkx
python-docx
